ARM: dts: sun8i: Add gpio-regulator used on Orange Pi Zero
This is the same diff as megous commit db8cf09e, but ported over to the
Orange Pi Zero - Armbian 5.27 dev for that board uses a DTB that looks
like it is just a copy of the Orange Pi One.

The Xunlong schematics for the Orange Pi Zero confirm the same details.