ARM: sun8i: v3s: add EMAC device node
As we have now the driver for the V3s EMAC, add the device node of it.

The MII pins of the V3s EMAC is not wired out, so only internal PHY can
be used.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>
(cherry picked from commit a9740d55a35343599a22ca2557ee3861a2be9d0d)