Dont compile kernel as THUMB2 - increase compatibility
This option was mainly needed to fix running the kernel under qemu.
I suspect there might have been some change in qemu, since the kernel
used to work when compiled with THUMB2.

There is theoretically a size increase without THUMB2, but a quick
and naive check showed that the compressed zImage was similar size
before and after this change.